Organizes download package
